Implemented 945 - http://temerity.us/community/forums/viewtopic.php?t=945
To implement the required plugins feature, we need for plpluginmgr to only
allow plplugin to connect while there are missing required plugins.  This
meant that only PluginMgrControlClient was allowed to connect to plpluginmgr
as it the means to install plugins.  First all subclasses of BaseMgrClient
have a clientID, this is currently set to the name of the class.  Second the
verifyConnection protocol was modified to send over the Pipeline version +
release and the clientID.  The server performs all the version checking, and
returns an OK message if the the connection is successful.  Or when there
are missing plugins only clientID of PluginMgrControlClient is allowed, all
others are rejected with the error message that plpluginmgr is in a not
ready state.  Once all the plugins are installed, plplugingmgr accepts all
connections.