Implemented 945 - http://temerity.us/community/forums/viewtopic.php?t=945
From the user's perspective plpluginmgr has a new option called "--bootstrap=(previous Pipeline root directory)" and plplugin has a new options called "--list-required"

The required plugins feature is to assist studios during Pipeline upgrades when the local studio plugins are not automatically transferred to pipeline-latest.  Only the Temerity plugins are included in Pipeline upgrades.  In order for Pipeline to properly function the local studio plugins must be installed prior to plmaster and plqueuemgr start ups.  With the new verifyConnection protocal that includes the clientID, we can block connection to plpluginmgr until all required plugins are installed.  When a plugin is installed, the required plugins GLUE file for the Vendor is updated.  The required plugins GLUE file is read at the startup of plpluginmgr to ensure all the plugin necessary are loaded, if not plpluginmgr only allows connections from plplugin.

During the initial use of this feature, a studio can use the "--bootstrap=(previous Pipeline root directory)" to bootstrap the required plugins GLUE files.  Using a previous Pipeline root directory, the plugins are verified and recorded in a GLUE per Vendor.  Temerity plugins are of course ignored during the bootstrap process since the rpm will supply the latest GLUE file, also the Temerity required plugins GLUE file is set to read only since studios should not install Temerity plugins.

Two response messages were added, PluginCountRsp and PluginListRequiredRsp.  PluginCountRsp is returned after a successful plugin install but there are other required-plugins that need to be installed and any unregistered plugins detected, it is a subclass of SuccessRsp since the install operation was successful but the overall state of plpluginmgr's plugins is incomplete.  PluginListRequiredRsp is return for any "--list-required" requests.

plplugin users can use the "--list-required" to get a list of the required plugins that need to be installed and the unregisted plugins detected during plpluginmgr startup.